moves for_each_meta.h and location_match_mode.h to top level
