+++ include/nigiri/location_match_mode.h
