+++ include/nigiri/query_generator/generator_settings.h
@@ -2,10 +2,10 @@
 
 #include "geo/box.h"
 
+#include "nigiri/location_match_mode.h"
 #include "nigiri/query_generator/transport_mode.h"
 #include "nigiri/routing/clasz_mask.h"
 #include "nigiri/routing/limits.h"
-#include "nigiri/routing/location_match_mode.h"
 #include "nigiri/routing/transfer_time_settings.h"
 #include "nigiri/timetable.h"
