+++ include/nigiri/routing/query.h
@@ -7,9 +7,9 @@
 
 #include "nigiri/common/interval.h"
 #include "nigiri/footpath.h"
+#include "nigiri/location_match_mode.h"
 #include "nigiri/routing/clasz_mask.h"
 #include "nigiri/routing/limits.h"
-#include "nigiri/routing/location_match_mode.h"
 #include "nigiri/routing/transfer_time_settings.h"
 #include "nigiri/td_footpath.h"
 #include "nigiri/types.h"
