+++ src/routing/get_fastest_direct.cc
@@ -3,8 +3,8 @@
 #include "utl/get_or_create.h"
 
 #include "nigiri/common/dial.h"
+#include "nigiri/for_each_meta.h"
 #include "nigiri/routing/dijkstra.h"
-#include "nigiri/routing/for_each_meta.h"
 #include "nigiri/special_stations.h"
 
 namespace nigiri::routing {
