+++ src/routing/raptor/reconstruct.cc
@@ -8,7 +8,7 @@
 #include "utl/overloaded.h"
 
 #include "nigiri/common/delta_t.h"
-#include "nigiri/routing/for_each_meta.h"
+#include "nigiri/for_each_meta.h"
 #include "nigiri/routing/journey.h"
 #include "nigiri/routing/raptor/raptor_state.h"
 #include "nigiri/rt/frun.h"
